==> 01_transportation_network_vietnam.ipynb <==
# Transportation Network of Vietnam
In this notebook, we construct a road network graph of Vietnam from the dataset.


# Imports
import osmnx as ox
import networkx as nx
import pickle

# Use Hanoi for testing
place_name = "Hanoi, Vietnam"
G_main = ox.graph_from_place(place_name, network_type="drive", simplify=True)

print(f"Directed road graph: {len(G_main.nodes())} nodes, {len(G_main.edges())} edges")

# Save graph for later use
with open("road_graph.pkl", "wb") as f:
    pickle.dump(G_main, f)
